New types for method arguments.

+++ constantDeclaration.cpp
@@ -8,7 +8,7 @@
 namespace CODEGEN {
 
 IR::Type *gen_constant_type() {
-    std::vector<int> percent = {80, 20};
+    std::vector<int64_t> percent = {80, 20};
     IR::Type *tp = nullptr;
     switch (randind(percent)) {
     case 0: {
